requirements.txt
appdirs==1.4.4
asgiref==3.3.0
black==20.8b1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-dirtyfields==1.4
django-filter==2.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
Markdown==3.3.3
mypy-extensions==0.4.3
newsapi-python==0.2.6
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
